+++ CPaaSSDK.framework/SubFrameworks/CPUtilities.framework/Headers/CPConstants.h
@@ -9,4 +9,4 @@
  *
  */
 
-#define CPaaS_SDK_VERSION @"2.12.0"
+#define CPaaS_SDK_VERSION @"2.13.0"